requirements.txt
pygame
python-vlc
Adafruit-CharLCD
rich
requests
RPi.GPIO
apscheduler